copy() CODE_ALT_ARRAY fix

+++ src/Input.cpp
@@ -349,7 +349,7 @@ void copy(TypeInput& in, TypeOutput* out, const TypeCode* type_code, const uint1
 			}
 			case CODE_ARRAY:
 			case CODE_ALT_ARRAY: {
-				size_t size = code[1];
+				uint16_t size = code[1];
 				if(code[0] == CODE_ALT_ARRAY) {
 					size = flip_bytes(size);
 				}
